+++ R/brace_linter.R
@@ -51,9 +51,8 @@ brace_linter <- function(allow_single_line = FALSE) {
       )
     ]")
 
-    lints <- c(lints, lapply(
+    lints <- c(lints, xml_nodes_to_lints(
       xml2::xml_find_all(source_expression$xml_parsed_content, xp_open_curly),
-      xml_nodes_to_lint,
       source_expression = source_expression,
       lint_message = paste(
         "Opening curly braces should never go on their own line and",
@@ -69,9 +68,8 @@ brace_linter <- function(allow_single_line = FALSE) {
       @col1 = { xp_open_preceding }/@col2 + 1
     ]")
 
-    lints <- c(lints, lapply(
+    lints <- c(lints, xml_nodes_to_lints(
       xml2::xml_find_all(source_expression$xml_parsed_content, xp_paren_brace),
-      xml_nodes_to_lint,
       source_expression = source_expression,
       lint_message = "There should be a space before an opening curly brace."
     ))
@@ -102,9 +100,8 @@ brace_linter <- function(allow_single_line = FALSE) {
       )
     ]")
 
-    lints <- c(lints, lapply(
+    lints <- c(lints, xml_nodes_to_lints(
       xml2::xml_find_all(source_expression$xml_parsed_content, xp_closed_curly),
-      xml_nodes_to_lint,
       source_expression = source_expression,
       lint_message = paste(
         "Closing curly-braces should always be on their own line,",
@@ -118,18 +115,16 @@ brace_linter <- function(allow_single_line = FALSE) {
     # TODO (AshesITR): if c_style_braces is TRUE, this needs to be @line2 + 1
     xp_else_same_line <- glue::glue("//ELSE[{xp_else_closed_curly} and @line1 != {xp_else_closed_curly}/@line2]")
 
-    lints <- c(lints, lapply(
+    lints <- c(lints, xml_nodes_to_lints(
       xml2::xml_find_all(source_expression$xml_parsed_content, xp_else_same_line),
-      xml_nodes_to_lint,
       source_expression = source_expression,
       lint_message = "`else` should come on the same line as the previous `}`."
     ))
 
     xp_function_brace <- "//expr[FUNCTION and @line1 != @line2 and not(expr[OP-LEFT-BRACE])]"
 
-    lints <- c(lints, lapply(
+    lints <- c(lints, xml_nodes_to_lints(
       xml2::xml_find_all(source_expression$xml_parsed_content, xp_function_brace),
-      xml_nodes_to_lint,
       source_expression = source_expression,
       lint_message = "Any function spanning multiple lines should use curly braces."
     ))
@@ -153,9 +148,8 @@ brace_linter <- function(allow_single_line = FALSE) {
     ]
     "
 
-    lints <- c(lints, lapply(
+    lints <- c(lints, xml_nodes_to_lints(
       xml2::xml_find_all(source_expression$xml_parsed_content, xp_if_else_match_brace),
-      xml_nodes_to_lint,
       source_expression = source_expression,
       lint_message = "Either both or neither branch in `if`/`else` should use curly braces."
     ))