+++ R/expect_identical_linter.R
@@ -62,15 +62,14 @@ expect_identical_linter <- function() {
     ]")
 
     bad_expr <- xml2::xml_find_all(xml, xpath)
-    return(lapply(
+    xml_nodes_to_lints(
       bad_expr,
-      xml_nodes_to_lint,
       source_expression = source_expression,
       lint_message = paste(
         "Use expect_identical(x, y) by default; resort to expect_equal() only when needed,",
         "e.g. when setting ignore_attr= or tolerance=."
       ),
       type = "warning"
-    ))
+    )
   })
 }