+++ R/nested_ifelse_linter.R
@@ -31,9 +31,8 @@ nested_ifelse_linter <- function() {
 
     bad_expr <- xml2::xml_find_all(xml, xpath)
 
-    return(lapply(
+    xml_nodes_to_lints(
       bad_expr,
-      xml_nodes_to_lint,
       source_expression = source_expression,
       lint_message = function(expr) {
         matched_call <- xml2::xml_text(xml2::xml_find_first(expr, "expr/SYMBOL_FUNCTION_CALL"))
@@ -41,7 +40,7 @@ nested_ifelse_linter <- function() {
         paste(lint_message, "instead, try (1) data.table::fcase; (2) dplyr::case_when; or (3) using a lookup table.")
       },
       type = "warning"
-    ))
+    )
   })
 }