+++ R/object_name_linters.R
@@ -65,13 +65,11 @@ object_name_linter <- function(styles = c("snake_case", "symbols")) {
 
     matches_a_style <- Reduce(`|`, style_matches)
 
-    lapply(
+    xml_nodes_to_lints(
       assignments[!matches_a_style],
-      xml_nodes_to_lint,
       source_expression,
       lint_message = lint_message,
-      type = "style",
-      global = TRUE
+      type = "style"
     )
   })
 }
@@ -187,13 +185,11 @@ object_length_linter <- function(length = 30L) {
 
     too_long <- nchar(nms_stripped) > length
 
-    lapply(
+    xml_nodes_to_lints(
       assignments[too_long],
-      xml_nodes_to_lint,
       source_expression = source_expression,
       lint_message = lint_message,
-      type = "style",
-      global = TRUE
+      type = "style"
     )
   })
 }